+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - Fixed script to update pdc with description
   [eikichi18]
+- Add getObjSize info in File field serializer.
+  [cekk]
 
 
 6.1.10 (2024-01-16)