+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 6.1.11 (unreleased)
 -------------------
 
+- Fixed script to update pdc with description
+  [eikichi18]
 - Add getObjSize info in File field serializer.
   [cekk]