+++ CHANGES.rst
@@ -7,6 +7,10 @@ Changelog
 - Added new indexer, catalog index and query operation for canale_digitale_link field of Servizio CT
 [deodorhunter]
 
+- Fixed script to update pdc with description
+  [eikichi18]
+- Add getObjSize info in File field serializer.
+  [cekk]
 
 
 6.1.10 (2024-01-16)