+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog
 - Add collective.volto.enhancedlinks dependency (needed for slate integration).
   [cekk]
 - Add enhancedlinks infos in File field serializer.
+- Fixed script to update pdc with description
+  [eikichi18]
+- Add getObjSize info in File field serializer.
   [cekk]